Adding L1000 cryostat+accessories (#28)
* Added L1000 cryostat+accessories

First attempt at an implementation of a LEGEND-1000 cryostat.

Four concentric polycones make up the most important part of
the cryostat volume - An outer steel cryostat, a vacuum gap,
an inner steel cryostat, and atmospheric liquid argon.

The reentrance tube and underground argon are also two con-
centric polycones, but with a different construct method
than the other polycones.

The support structure holding the cryostat ("the skirt")
and the steel beam supporting the skirt (the "foot") have
been added in the water tank.

The PMMA neutron moderating shield for LEGEND-1000 has
been implemented both as a single G4Polyhedra ("simple")
and using STL files from the Padova group ("stl"), with
of course the option to "omit". This should not break
legacy code, because the implementation allows for the
neutron shield to omit by default if unspecified.

This commit includes:

 - A completely reworked cryostat file (cryo.py)
 - A minorly updated special_metadata.yaml (includes
 the moderator as an option)
 -

* See last commit message.

* See last commit message.

* Added stl files for neutron moderator.

* Fixed bug where skirt/foot z displacement was incorrect without the inclusion of the water tank. Edited special metadata file to be more consistent with expectations. Added type hinting to new cryo.py functions.

---------

Co-authored-by: Clay Barton <cbarton@lfs2.mpi-hd.mpg.de>